Introduce `HandleMessageAdvice` marker interceptor to cover the case when an `Advice` can be advices as well. Remove unused `setAdviceChainIfPresent()` method in the `AbstractMethodAnnotationPostProcessor`
